+++ app.js
@@ -1,0 +1,24 @@
+import express from 'express';
+import morgan from 'morgan';
+import helmet from 'helmet';
+import cookieParser from 'cookie-parser';
+import bodyParser from 'body-parser';
+
+import globalRouter from './routers/globalRouter';
+import userRouter from './routers/userRouter';
+import videoRouter from './routers/videoRouter';
+
+const app = express();
+
+// The order matters
+app.use(cookieParser());
+app.use(bodyParser.json());
+app.use(bodyParser.urlencoded({ extended: true }));
+app.use(helmet());
+app.use(morgan('dev'));
+
+app.use('/', globalRouter);
+app.use('/user', userRouter);
+app.use('/video', videoRouter);
+
+export default app;
