+++ package.json
@@ -6,7 +6,7 @@
   "author": "jerome",
   "license": "MIT",
   "scripts": {
-    "start": "nodemon --exec babel-node index.js --delay 2"
+    "start": "nodemon --exec babel-node init.js --delay 2"
   },
   "dependencies": {
     "@babel/core": "^7.5.4",
