+++ routers/globalRouter.js
@@ -1,0 +1,5 @@
+import express from 'express';
+
+const globalRouter = express.Router();
+
+export default globalRouter;
