+++ routers/userRouter.js
@@ -1,0 +1,5 @@
+import express from 'express';
+
+const userRouter = express.Router();
+
+export default userRouter;
