+++ routers/videoRouter.js
@@ -1,0 +1,5 @@
+import express from 'express';
+
+const videoRouter = express.Router();
+
+export default videoRouter;
